+++ tests/test.nightwatch.conf.js
@@ -1,7 +1,10 @@
 const nightwatchConfig = require('../lib/nightwatch/nightwatch.config.js').default;
 const webpackConfig = require('./test.config.js');
+const path = require('path');
+
+const testPath = path.join('tests', 'nightwatch');
 
 // Allows one to specify port number in configuration
-const config = nightwatchConfig(webpackConfig, ['tests/nightwatch/'], 9000);
+const config = nightwatchConfig(webpackConfig, [testPath], 9000);
 
 module.exports = config;