+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Use path.join to allow for windows development
+* Removed sauce labs config and install java 8 in a different way for travis ci
 
 2.1.0 - (October 5, 2017)
 ------------------